+++ backend/framework/sdk/src/test/java/io/metersphere/sdk/base/param/EnumValueParamGenerator.java
@@ -1,4 +1,4 @@
-package base.param;
+package io.metersphere.sdk.base.param;
 
 
 import io.metersphere.sdk.constants.ValueEnum;
