+++ backend/services/system-setting/src/test/java/io/metersphere/system/controller/CommonParamTests.java
@@ -1,6 +1,6 @@
 package io.metersphere.system.controller;
 
-import base.BaseTest;
+import io.metersphere.sdk.base.BaseTest;
 import io.metersphere.system.controller.param.BasePageRequestDefinition;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.MethodOrderer;
