+++ backend/services/system-setting/src/test/java/io/metersphere/system/controller/PluginControllerTests.java
@@ -1,6 +1,6 @@
 package io.metersphere.system.controller;
 
-import base.BaseTest;
+import io.metersphere.sdk.base.BaseTest;
 import io.metersphere.sdk.constants.PermissionConstants;
 import io.metersphere.system.domain.Plugin;
 import io.metersphere.system.mapper.PluginMapper;
