+++ backend/services/system-setting/src/test/java/io/metersphere/system/controller/user/UserControllerTests.java
@@ -1,6 +1,6 @@
 package io.metersphere.system.controller.user;
 
-import base.BaseTest;
+import io.metersphere.sdk.base.BaseTest;
 import io.metersphere.sdk.controller.handler.ResultHolder;
 import io.metersphere.sdk.dto.BasePageRequest;
 import io.metersphere.sdk.dto.ExcelParseDTO;
